test: migrate ci_context.rs unit tests to integration tests
Migrated all 8 tests from src/ci/ci_context.rs to tests/integration/ci_context_unit.rs. Converted git2 API calls to CLI git commands and made get_rebased_commits method public for testing.
